Update all Brew Formulas to have SHA256 Hashes
Brew is depricating sha1 hashes, so this updates all the hashes
in these brews to SHA256. I manually downloaded these brew files
and confirmed their SHA1 mathces what we have, then I got a SHA256
for the package and replaced the SHA1.
The only exception is Zlib. Which the downloaded SHA1 didbn't match
what the Formula says. But I found a newer Formula, for the same version,
that has a matching SHA256.